requirements.txt
django >= 2.0
djangorestframework >= 3.0
django-debug-toolbar
jinja2
pytest
pytest-django
pytest-coverage
